tests are working.. but doesn't align with the commented requirement.
